+++ src/containerapp/HISTORY.rst
@@ -9,6 +9,7 @@ Upcoming
 * 'az containerapp service': add support for creation and deletion of MariaDB
 * 'az containerapp create/list': support --environment-type parameter
 * 'az containerapp logs show': fix raising error for response status code is not OK
+* 'az containerapp auth show/update': support api-version 2023-05-02-preview
 * 'az containerapp create': --yaml support property additionalPortMappings for api-version 2023-05-02-preview
 
 0.3.36